requirements.txt
openai>=1.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
mcp>=1.2.0
